Move common tests to trait
